change domain/co-domain of the account relation (#5)
This PR changes the domain/co-domain of the relation. Previously, the
account relation is a production of a finite set of `address`, the
finite set is all addresses stored in the `account-graph`. The relation
is stored as a directed adjacency graph. In this approach, the main
concern is storage efficiency, for each relationship or edge, storage of
two nodes with two list is required, also a counter of incoming edges is
needed for each node in order to decide a node can be removed or not.

This PR changes the domain/co-domain of the relation to the set of all
`address`, `account-graph` stores all edges, account properties,
relationship properties, it avoids maintaining a set of `addresses` as
domain/co-domain.